% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_HMM_CORES}
\alias{DEFAULT_HMM_CORES}
\title{Default HMM Core Count}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_HMM_CORES
}
\description{
Default number of CPU cores for HMM per-cell parallelism.
}
\keyword{internal}
