% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HETEROZYGOSITY_THRESHOLD}
\alias{HETEROZYGOSITY_THRESHOLD}
\title{Heterozygosity Threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
HETEROZYGOSITY_THRESHOLD
}
\description{
MAF threshold for het/hom classification.
}
\keyword{internal}
