% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HMM_DEFAULTS}
\alias{HMM_DEFAULTS}
\title{HMM Default Parameters}
\format{
An object of class \code{list} of length 4.
}
\usage{
HMM_DEFAULTS
}
\description{
Default parameters for HMM algorithm tuning.
}
\keyword{internal}
