% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_per_chromosome.R
\name{hmm_per_chromosome}
\alias{hmm_per_chromosome}
\title{Per-Chromosome HMM Implementation}
\description{
Functions for fitting HMM independently to each chromosome and
  parallel multi-cell processing.
}
