% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_norm}
\alias{log_norm}
\title{Log-Density of Normal Distribution}
\usage{
log_norm(x, mu, var)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{mu}{Numeric, mean parameter.}

\item{var}{Numeric, variance parameter.}
}
\value{
Numeric vector of log-density values.
}
\description{
Calculates the log-density of a normal distribution.
}
\keyword{internal}
