% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logsumexp}
\alias{logsumexp}
\title{Log-Sum-Exp Function}
\usage{
logsumexp(v)
}
\arguments{
\item{v}{Numeric vector of log values.}
}
\value{
Numeric, the log of the sum of exponentials.
}
\description{
Computes the log-sum-exp in a numerically stable way.
}
\keyword{internal}
