% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{LOW_HET_RATE_THRESHOLD}
\alias{LOW_HET_RATE_THRESHOLD}
\title{Low Heterozygosity Rate Threshold}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LOW_HET_RATE_THRESHOLD
}
\description{
Threshold for considering a sample "low heterozygosity" when deciding
  whether to apply depth-anchored HMM overrides.
}
\keyword{internal}
