% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MAX_SIGMA_MAF}
\alias{MAX_SIGMA_MAF}
\title{Maximum MAF Sigma}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MAX_SIGMA_MAF
}
\description{
Upper bound for MAF standard deviation.
}
\keyword{internal}
