% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MIN_SIGMA_MAF}
\alias{MIN_SIGMA_MAF}
\title{Minimum MAF Sigma}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MIN_SIGMA_MAF
}
\description{
Lower bound for MAF standard deviation.
}
\keyword{internal}
